pyproject.toml
[tool.poetry]
name = "odin-backend"
version = "0.1.0"
description = "Backend para gerenciamento de processos de dados escolares"
authors = ["Samuel <colaco.samuel8@gmail.com>"]
readme = "README.md"
package-mode = false

packages = [
    { include = "src", from = "." }
]

[tool.poetry.scripts]
dev = "python src/server.py"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0" }
pydantic = "^2.5.0"
motor = "^3.3.2"
pymongo = "^4.6.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
dependency-injector = "^4.41.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[virtualenvs]
in-project = true